Move IconButton to be a proper React Common component
