+++ Dnn.AdminExperience/ClientSide/Bundle.Web/src/dnn-common-components.js
@@ -18,6 +18,7 @@ import {
     FolderPicker,
     GridCell,
     GridSystem,
+    IconButton,
     InputGroup,
     Label,
     Modal,
@@ -69,6 +70,7 @@ export {
     FolderPicker,
     GridCell,
     GridSystem,
+    IconButton,
     InputGroup,
     Label,
     Modal,
