+++ Dnn.AdminExperience/ClientSide/Dnn.React.Common/src/PermissionGrid/GridRow.jsx
@@ -2,7 +2,7 @@ import React, { Component } from "react";
 import PropTypes from "prop-types";
 import GridCell from "../GridCell";
 import StatusSwitch from "./StatusSwitch";
-import IconButton from "./IconButton";
+import IconButton from "../IconButton";
 
 class GridRow extends Component {
     constructor(props) {
