--- Dnn.AdminExperience/ClientSide/Dnn.React.Common/src/PermissionGrid/IconButton/style.less
@@ -1,27 +1,0 @@
-@import "../../../styles/index";
-.icon-button{
-    display: inline-block;
-    svg{
-        vertical-align: top;
-        fill: @alto;
-    }
-
-    &:hover{
-        svg{
-            fill: @thunder;
-        }
-    }
-
-    &:active, &.activated{
-        svg{
-            fill: @curiousBlue;
-        }
-    }
-}
-.icon-flat{
-    display: inline-block;
-    svg{
-        vertical-align: top;
-        fill: @alto;
-    }
-}
